Started building the core back up
I've got the tests in ScoringEngine.WebApi.test.tests_organizations.OrganizationsList working. The endpoints in endpoints.py for Users needs to be updated to match what's been added in the Organization object. The only difference is a few small tweaks with the serialze() method, the addition of the create() method, and the tweaks to the __init__() method (with response ot the newly added create() method). This commit has something like 4 or 5 tests passing out of 91ish.