Implemented ModelWrapper and POST data bug
I implemented ModelWrapper which cut out a ton of repeated lines in endpoints.py. This also fixed a bunch of random issues here and there. It's only been applied to subclasses of Compeititon. Plugin, Organizations and User hasn't be changed to use it yet.
I also fixed a SUPER annoying issue with a few of the create**** functions in utils for the test cases. The example post data dictionary was being updated with specific IDs for related database entries. Updating that dictionary returned None, where I thought it would return the updated dict. This has been fixed and totally works now. That would have been a satisfying issue to close had I made one for it...