Starting ScoringEngine module support
Created a model for registering scoring engines. Models are created at start. I've written a little bit of stuff to import the scoring engine based off the model for it in the database, then run the run() function in the engines __init__() class. Now to write the rest of the surrounding framework and make it so the engine can be associated with competitions.