Finished busy word in client.py
Finished most of the busyword stuff in client.py. I'll have to go back and tweak some of it in the future, but I got that working. Moving development over to a proper linux box, so I'm pushing all the super incomplete changes to github (easier to just reclone the repo to a vm).