Cleanup and fixes
Added some basic exception handling to the api methods, that way errors are reported through celery cleanly. I updated the client so that it handles those stack traces nicely.

I moded the asDict and edit methods of all of the modelwrapper classes to the ModelWrapper base class. This required adding a list called 'fields', so that the base class knows what attributes to look for. That was an awesome improvement!