Changed daemon provider
I've copied a common daemon class found online, since it provides some nice quality of life improvements over the old daemon I was using. With this, I've added a status option to see what the status of the daemon is.

Additionally, I've made progress on debugging the duplicate messages received by the client: the number of times you have to run 'organization-get' (or any *-get task) for data to be duplicated is the same as the number of workers that are spawned. By default, there were 4 workers, so the command would have to be run 4 times. By limiting the workers to 2, you only have to rut it twice to get duplicate data.