Refactory updates to WebInterface
Commiting and pushing early since my dev VM is acting a little weird.
The WebInterface is using a slightly modified version of the CLI client that is in the Cssef directory. It's mostly working, and pages just need to be updated to work with the new method of querying the api and preparing the page context.