Created CssefClient
I had two copies of client.py for the WebInterface and Cssef, since importing it accross the folders was difficult. I decided to create a formal package for the client, that way I can install it and import it from the WebInterface, while still using it for testing in Cssef. I've written it as a proper python package, so you just need to run 'python setup.py install'. See the README.md for more info.
I did some other housekeeping like deleting the 'start_cssef' since it was outdated. Updated the .gitignore and removed the old serializers.py file from the Cssef folder.