Packaged cssefserver
The cssef server can now be installed via pip. The README.md for that directory has been updated to describe the installation process. I did notice that celery will fail to run on the server because it is running as root, in which case C_FORCE_ROOT need to be true. I haven't had a chance to properly document this.