Reorganized site structure
Hugely reorganized the way the site is structured to make management easier. Each componenet of the site is in its own module under the 'modules' directory. This will especially make future expansion easier.
I've also built in a debug view for every page. That used to just exist for the createEditUser and creatEditOrganization pages, but I found it super useful, so I've added it to the base page template. As of now, it's enabled everywhere because the default 'debug' value in the BaseContext object is set to True. This will need to be changed to pull from values in the 'cssefd.conf' file eventually.