Fixed unnecessary api hit issue
There was an issue where the form for the user creation page was making an api call each time it was referenced in the html template for that page. This is because the __init__ method made an api call to populate the list of available organizations. This was a problem because the usercreation page referenced the form 16 (last time I counted) times, which meant the api would be queried 16 times through the process of rendering the page. It turns out this was because the form being passed to the template wasn't initalized. Here I've changed the getContext method for the FormContext object to initialize the form before adding it to the context.