More WebInterface fixes
Fixed edit pages after the form api hit issue in the previous commit. I've changed URLs to use IDs now rather than specified URLs. This makes handling urls a lot easier and is 1 less field to track in the database. I haven't actually removed the URL option from the database,server,client, or webinterface forms. It is simply not used. Lastly, the server can include fields that don't have explicit properties in the modelwrapper class. this means if there's something in the 'fields' list and the dictionary is being prepared, the ModelWrapper base class won't error out, but will rather try to fiend that attribute in the model class.