Last major restructure of server components
I've moved everything around in the server to split everything up into either the framework or modules. It is much easier to tell what code is for a module and what is part of the actual framework. Each 'module' is a major component to the server. Currently the available modules are 'competitions', 'organization', 'user' and 'documents' (though Documents hasn't been updated or maintained in a very long time. Stuf *should* work as it is now- though it was only tested with a pre-existing database, so I'm not sure if the database will be correctly generated the first time...