User + Organization module restructuring
I created a new module called 'account' and moved all of the User and Organization endpoints under that. This makes more logical sense, and will improve the organization and user classes working together. Additionally, I've resolved issue #34, so there is no longer a createMember function in the organization. I have not caried this change over to the competition class, since that will require considerably more work to be as dynamic as I would prefer.