Implementing authentication
I've implemented authentication from the server side. The client authentication is hardcoded right now, because I want to make sure I completely nail down the way the server should work. I had to do some module rearrangement to make authentication work smoothly.
Now I need to set up the initial administrator organization and user, which can't be done since the client is completely locked out of the system w/o the admin user. We can set an 'administrator token' in the config file, which is just a long string of text (not an actual token) which can be used to "authenticate" the first few client requests to set up orgs and users.
To do this though, I need to make the server actually read the configuration file...