Code cleanup
Authorization checking on api endpoints is done inside the try/except blocks, so any unhandled issues will be somewhat gracefully handled. The framework.__init__ was cleaned up to, in preparation for improving frequent database and configuration instantiation problems. References to database or configuration imports/instantiations have been changed to import from the framework (yay!).

The client will now pull the organization and username to login as from the configuration, so now you only have to provide --password, whereas before you'd also have to provide --username and --organization (so much typing!!). On top of this, if you do not want to print the password on the command line, you can provide no arguments, and it will prompt you for the password. So cool!