Implemented framework plugins
You can import a modules that consumes the framework by listing it in the `installed-plugins` section of the server configuration file. It will be imported when the project starts, and any available tasks and endpoints will be registered. Clients can then call the endpoints. I've created a new module called 'cssef-competition-attack-defend' as a module to score CCDC-like competitions (note that this will not be the final module name since it's super lame right now). There are still several huge issues to solve now, but the core functionality is there.
Also note that I've refered to these modules simply as 'plugins'. These are different than the plugins I've refered to in past commits, in which I was making modules to score specific services defined within what is now the cssef-competition-attack-defend module. Those "plugins" will be given a different name so at to avoid confusion- but that's for a later commit :)