More restructuring
Cleaned up the structure of cssefserver. This is necessary to fix the database handling (both for modules and connections). Most stuff should be updated to work with this new structure, however it has not been tested. I also cleaned up older code in some places (example: the methods in taskutils.py), which should hopefully increase efficiency