Finished implementing databasing for plugins
Plugins can now define database tables and models to be created in the database. This required a bit more module restructuring, and small tweaks to database code in several places. Several issues with the API in general have been brought to my attention- largely regarding the dependancies required to create instances of basic objects (like Users and Organizations). I'll be making a more concious effort to clean these dependancies up to make the API more accessible.