Slight client cleanup
The work I did to implement the authentication was somewhat messy from the client side. I've started cleaning up the classes and whatnot. I've consolidated them a little and moved some of them from cssef-cli to cssefclient. I realized this was all super necessary since it is required for the WebInterface to properly interact with the server now.