Partially completes #29
Took care of some very badly needed restructuring on the client. The cssef-cli and its configuration file are no longer actually in the package, and the main file has been split up to improve imports. While I'm not properly catching ctrl+c or anything like that, I've added 'expires=10' to the start_task call in CeleryEndpoint which will mean the task is not executed if it's picked up by a worker after 10 seconds.