Replaced celery with flask + rpc
Celery had a few super serious shortcomings with calling tasks and whatnot. I've completely removed celery, and replaced it an rpc implementation using flask. In retrospect, this seriously simplifies installation since we are no longer relying on rabbitmq. There are still a few rough corners that need to be smoothed out, but for the most part everything still works- that was a crazy 3 days of research, patching and debugging...