Small update to unit tests
Now that the the __init__ method of cssefserver has been cleaned up, I can properly starting updating unit tests for it. The old tests have just been deleted, and I've started writing new ones. There are still a few smaller issues to work out with the tests, but this is a good start.

I also updated the travis.yml file to stop running tests on WebAPI and WebUI, since the WebAPI has been abandonded for now, and the WebUI is super old.