House keeping. Doc fixes/updates + pylint fixes
I've tried to fix a few issues with getting unit tests working, as well as autodocs running for RTD. I decided to run pylint over the server and client code to start cleaning that up a little. It looks like I've resolved a bunch of smaller issues that haven't been issues yet due to lack of testing.