Fix permissions issues with config files
There were a lot of permissions issues while running the daemon. It needs to be run with root or a user with write permissions to /etc/cssef as well as /var/log/cssef. It will not automatically set permissions for those directories, but it should eventually. Making another card to get that fixed...