Using adminlte for webinterface frontend
Updated a few of the templates that I can actually view right now. I also made a few small bug fixes to the client and server. I just saw that the adminlte content contains a TON of pngs and whatnot, so please forgive me for commiting all of this. I'll sort out a better way of including it later. I also cleaned some of the setup.py files, and removed the old MANIFEST and LICENSE files that weren't exactly necessary.