Fixed pip package installs
Added a git precommit script to be run before committing. It will make sure the packages build and install, and that the unit tests for both will run correctly. The unit tests don't pass right now though, so committing will have to be done with the --no-verify option... The pre-commit script can also be used for testing in general. By default it will install the pacakges into a virtual environment, but you can test on your system by providing the --no-virtual-env option.