Mostly fixed client
rpc requests to the server are working again. Fixed a bunch of problems with the various client features: endpoint caching and authentication for instance. Token authentication is still technically broken- the client will correctly authenticate with tokens, but the auth dictionary isn't being built in the correct order, so it's super broken. I'm going to have to do some tricky refactoring to fix this for the cli.

Few random fixes for other stuff like the pre-commit checks.