Updated templates
I'm getting ready to consolidate a bunch of the templates. The frontend "library" I'm using is fantastic and makes it easier to increase the functionality of each page. I've started to update some of the pages to take advantage of this functionality, though it's been slow. I've had to hack away at the Context class a little to make it fit this newer approach. I proper redesign/update to that will be needed very soon.