Drastically improved endpoint loading
Cleaned up the way endpoints are tracked and loaded by the server. This integrates very nicely with the plugin class I made a commit or two ago. Updated the only plugin package I've got to work with this, and confirmed that it's all working here. The client fails because the data structure that is sent by AvailableEndpoints has changed, so the client will need to be fixed- but the server side is working and that's what matters!