Updated client to handle new endpoint system
The client works with the new way of reporting available endpoints. The simplified endpoint list really cut down on the overhead necessary for the client. There were a number of small fixes throughout the rest of the project that were necessary, but most of it should kind of be working now. calling 'cssef-cli competition get' seems to route to a serverside rpc endpoint at least, so that's working (I didn't fully test it because of more immediate problems).