Implemented server runtime configurations
This was a feature that was described in the documentation that wasn't actually implemented (I'm really sorry!). I've implemented it here and confirmed it's working. The only case it may not work is for the installed-plugins option which will require some additional parsing. If you provide the runtime configuration of --pidfile, you will need to provide that same configuration for subseqent commands to apply to that instance of the server.