Cleaned up cssefserver.tasks
Moved the RewnewToken and Login tasks to cssefserver.account.tasks, because that's where they belong. Cleaned up a lot of code that was written specifically to support those two classes. Improved logging in several places including when the database fails to initialize (thus killing the service). Updated old commit reference in modelbase. Changed several old references to Celery, and updated some of the server documentation.