Removed WebApi
Finally cutting off the dead limb here. The WebApi was a remnant of the previous server/client model. Since the project has changed and is now using RPC calls built into the server process, we've kind of come full circle but don't need the legacy code. Not only will it be saved in versioning history, but also in our hearts.

So long, and thanks for all the fish.