Improved authentication object management
The concept of an "auth dict" was not well implemented, and caused lot of other small problems in the code base. I've written a proper class to build a working authentication dictionary based on the provided authentication information (username/org/password/token/admin token). That class is now in the cssef-client library (it wasn't before). The authentication token is also loaded earlier now, making the auth buildup process easier.

Lastly, auth token renewal will no longer be attempted when the user is not using a token for authentication.