Updated client config file + docs
The client config file didn't have anything useful in it, and default connection settings were relying on hardcoded values. I'm trying to move away from that, so I fleshed out and updated the config file. Also updated a lot of the client config documentation. I removed the configuration for the rpc-base-uri (can't remember the exact config name) since it will likely never be needed.

Made a few small tweaks to the server documentation and makefile.