Client improvements
I'm trying to streamline the client a little more, so I've trying to shift as much functionality from the cssef-cli to the CssefClient class as possible- this will be needed for when the webui starts consuming CssefClient. Authentication has been moved to the AuthBuilder class as much as possible. Locally implemented tasks aren't working right now, so calling 'user login' results in the token being printed to the screen. I will need to write a way to attach a class to a remote endpoint...