Added/improved docstrings + sphinx config tweak
There was a small error with the sphinx config that broke local documenation builds (recursion error). I added a small check to prevent that, and it *should* have no effect on the rtd build...

Added a ton of docstrings now that apidoc is working. Going to sleep for now, but I'll continue adding documentation tomorrow.