Removed get_empty_return_dict()
Finally got around to replacing the abomination that was get_empty_return_dict()- the result of trying to program too late at night with questionable sobriety. Replaced it with a class similar to the CommandOutput class in the client: EndpointOutput. CssefRCPEndpoint.on_request() should now *always* return an instance of EndpointOutput.