cssefserver.Configuration cleanup and unit tests
Added unit tests for cssefserver.Configuration, and have a solid approach to how to implement unit tests from here out. That will be the main focus for a little while. I had started trying to clean up Configuration, and wall as CssefRPCEndpoint, but it's been messy and very difficult to tell if stuff is still working afterward. Updated issue #22 to track unittest progress.