RPCEndpoint server reference + endpoint loading
RPCEndpoint now holds a reference to the server object that owns it (like what is described here: http://stackoverflow.com/questions/10791588/getting-container-parent-object-from-within-python), that way we can move stuff from Configuration that really should be in CssefServer. I've already done this with endpoint_sources. I split up the method that was loading endpoints in CssefServer. It's several functions now, which will hopefully make it easier to deal with importing the endpoint source information.