Logging with logger
Removed journal dependancy, because it was unnecessary. Logging using logger. Re-arranged the script that starts the server such that imput parsing is available through the library. ModleWrapper objects now have access to the server. This might be dangerous, but it makes it much easier to access the configuration and database, which extends usability.